refactor: rename interface files and update related imports
- Renamed interface files for better naming consistency
- Updated all imports to reflect new interface filenames
