+++ backend/nestjs/src/modules/email/interfaces/IService.ts
