+++ backend/nestjs/src/modules/product/interfaces/IService.ts
@@ -1,0 +1,6 @@
+import { SearchDto } from "../dto";
+import { SearchResponse } from "../types";
+
+export interface IProductService {
+  search(query: SearchDto): Promise<SearchResponse>;
+}
