+++ backend/nestjs/src/modules/product/interfaces/services/IS3Service.ts
@@ -1,10 +1,3 @@
-import { SearchDto } from "../dto";
-import { SearchResponse } from "../types";
-
-export interface IProductService {
-  search(query: SearchDto): Promise<SearchResponse>;
-}
-
 export interface IS3Service {
   putObject(
     id: string,
