+++ backend/nestjs/src/modules/shop/interfaces/IService.ts
